# Server/swagger.yaml
swagger: "2.0"
info:
  description: "API của Website Hỗ trợ đào tạo SGU"
  version: "1.0.0"
  title: "Website Hỗ trợ đào tạo SGU"
  contact:
    email: ""
host: ""
...
basePath: "/api"
schemes:
- "http"
- "https"
...
paths:
...
  /admin/quyen-tai-khoan/DanhSachQuyenTK:
    get: 
      tags:
        - "QuyenTaiKhoan"
      summary: Admin lấy danh sách quyền tài khoản
      produces:
        - application/json     
      parameters:
      - in: query
        name: page
        type: integer
        required: true
        description: số trang (ví dụ trang 1, trang 2), giá trị > 0
      - in: query
        name: pageSize
        type: integer
        required: true
        description: giới hạn số lượng 1 trang, giá trị > 0
      - in: query
        name: keyword
        type: string
        required: true
        description: mã quyền tài khoản hoặc tên quyền tài khoản
      responses:
        200: 
          description: Lấy danh sách quyền tài khoản thành công
          schema:
            type: object
            example: {
              "status": true,
              "message": "Lấy danh sách quyền tài khoản thành công",
            }
        400:
          description: Lấy danh sách quyền tài khoản thất bại
          schema:
            type: object
            example: {
              "status": false,
              "message": "Lấy danh sách quyền tài khoản thất bại",
            }
      security:
      - ApiKeyAuth: []
  /admin/quyen-tai-khoan/ChiTietQuyenTK/{MaQTK}:
    get:
      tags:
        - "QuyenTaiKhoan"
      summary: "Chi tiết thông tin quyền tài khoản"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaQTK
          type: string
          required: true
          description: Mã quyền tài khoản
      responses:
        200:
          description: Chi tiết thông tin quyền tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chi tiết thông tin quyền tài khoản thành công.",
            }
        400:
          description: Chi tiết thông tin quyền tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Chi tiết thông tin quyền tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: []
  /admin/quyen-tai-khoan/Them:
    post: 
      tags:
        - "QuyenTaiKhoan"
      summary: Thêm quyền tài khoản
      produces:
        - application/json     
      parameters:
        - in: body
          name: MaQTK
          type: string
          required: true
          description: Mã quyền tài khoản
        - in: body
          name: TenQuyenTK
          type: string
          required: true
          description: Tên quyền tài khoản
        - in: body
          name: MaCN
          type: string
          description: Mảng các mã chức năng con được cách nhau bởi dấu ";". Ví dụ (CN;TK;CHNG)
        - in: body
          name: ChucNangCon
          type: string
          description: Mảng các tên chức năng con được cách nhau bởi dấu ";" ứng với mã chức năng, các tên cách nhau bởi dấu ",". Ví dụ (Thêm,Sửa,Xóa;Sửa,Xóa;Xem danh sách)
      responses:
        200:
          description: Thêm quyền tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Thêm quyền tài khoản thành công.",
            }          
        400:
          description: Thêm quyền tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message": "Thêm quyền tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: [] 
  /admin/quyen-tai-khoan/ChinhSua/{MaQTK}:
    put: 
      tags:
        - "QuyenTaiKhoan"
      summary: Chỉnh sửa quyền tài khoản
      produces:
        - application/json     
      parameters:
        - in: path
          name: MaQTK
          type: string
          required: true
          description: Mã quyền tài khoản
        - in: body
          name: TenQuyenTK
          type: string
          required: true
          description: Tên quyền tài khoản
        - in: body
          name: MaCN
          type: string
          description: Mảng các mã chức năng con được cách nhau bởi dấu ";". Ví dụ (CN;TK;CHNG)
        - in: body
          name: ChucNangCon
          type: string
          description: Mảng các tên chức năng con được cách nhau bởi dấu ";" ứng với mã chức năng, các tên cách nhau bởi dấu ",". Ví dụ (Thêm,Sửa,Xóa;Sửa,Xóa;Xem danh sách)
      responses:
        200:
          description: Chỉnh sửa quyền tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chỉnh sửa quyền tài khoản thành công.",
            }          
        400:
          description: Chỉnh sửa quyền tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message": "Chỉnh sửa quyền tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: []     
  /admin/quyen-tai-khoan/Xoa/{MaQTK}:
    delete:
      tags:
        - "QuyenTaiKhoan"
      summary: "Xóa thông tin quyền tài khoản"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaQTK
          type: string
          required: true
          description: Mã quyền tài khoản
      responses:
        200:
          description: Xóa thông tin quyền tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Xóa thông tin quyền tài khoản thành công.",
            }
        400:
          description: Xóa thông tin quyền tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Xóa thông tin quyền tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
  /admin/tai-khoan/Them:
    post: 
      tags:
        - "TaiKhoan"
      summary: Thêm tài khoản
      produces:
        - application/json     
      parameters:
        - in: body
          name: MaTK
          type: string
          required: true
          description: Mã tài khoản
        - in: body
          name: MaQTK
          type: string
          required: true
          description: Mã quyền tài khoản
        - in: body
          name: TenDangNhap
          type: string
          required: true
          description: Tên đăng nhập
        - in: body
          name: MatKhau
          type: string
          description: Mật khẩu đăng nhập
      responses:
        200:
          description: Thêm tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Thêm tài khoản thành công.",
            }          
        400:
          description: Thêm tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message": "Thêm tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: [] 
  /admin/tai-khoan/ChinhSua/{MaTK}:
    put: 
      tags:
        - "TaiKhoan"
      summary: Chỉnh sửa tài khoản
      produces:
        - application/json     
      parameters:
        - in: path
          name: MaTK
          type: string
          required: true
          description: Mã tài khoản
        - in: body
          name: MaQTK
          type: string
          required: true
          description: Mã quyền tài khoản
        - in: body
          name: TrangThai
          type: string
          required: true
          enum: ['Chưa kích hoạt','Đã kích hoạt']
          description: Trạng thái tài khoản
      responses:
        200:
          description: Chỉnh sửa tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Chỉnh sửa tài khoản thành công.",
            }          
        400:
          description: Chỉnh sửa tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message": "Chỉnh sửa tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: []     
  /admin/tai-khoan/Xoa/{MaTK}:
    delete:
      tags:
        - "TaiKhoan"
      summary: "Xóa thông tin tài khoản"
      produces:
        - application/json
      parameters:
        - in: path
          name: MaTK
          type: string
          required: true
          description: Mã tài khoản
      responses:
        200:
          description: Xóa thông tin tài khoản thành công.
          schema:
            type: object
            example: {
              "status": true,
              "message": "Xóa thông tin tài khoản thành công.",
            }
        400:
          description: Xóa thông tin tài khoản thất bại.
          schema:
            type: object
            example: {
              "status": false,
              "message" : "Xóa thông tin tài khoản thất bại.",
            }
      security:
        - ApiKeyAuth: []
...
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: headers
    name: x-access-token